+++ src/mqtt.c
@@ -102,7 +102,7 @@ bool naos_unsubscribe(const char *topic, naos_scope_t scope) {
   return ret;
 }
 
-bool naos_publish_raw(const char *topic, void *payload, uint16_t len, int qos, bool retained, naos_scope_t scope) {
+bool naos_publish_raw(const char *topic, void *payload, size_t len, int qos, bool retained, naos_scope_t scope) {
   // add base topic if scope is local
   if (scope == NAOS_LOCAL) {
     topic = naos_mqtt_with_base_topic(topic);
@@ -123,7 +123,7 @@ bool naos_publish_raw(const char *topic, void *payload, uint16_t len, int qos, b
 }
 
 bool naos_publish(const char *topic, const char *str, int qos, bool retained, naos_scope_t scope) {
-  return naos_publish_raw(topic, (char *)str, (uint16_t)strlen(str), qos, retained, scope);
+  return naos_publish_raw(topic, (char *)str, strlen(str), qos, retained, scope);
 }
 
 void naos_mqtt_stop() {
